invoker 功能优化 以及 checkTimeoutTak

+++ walle-common/src/main/java/cn/pany/walle/common/constants/NettyConstant.java
@@ -24,5 +24,5 @@ public final class NettyConstant {
     public static final String INTERFACE_CLASS_KEY ="interfaceClass";
 
 
-    public static final int DEFAULT_TIMEOUT = 10;
+    public static final Long DEFAULT_TIMEOUT = 10L;
 }
